Created new file for containing plotters
